app/requirements.txt
Django==5.*
gunicorn==22.*
psycopg[binary]==3.*

celery[redis]==5.*
redis==5.*

docling==2.*

sentence-transformers==5.*

pgvector==0.*

python-hwpx==2.*
pyhwp==0.1b15
djangorestframework==3.15.*
drf-yasg==1.21.*
pytest==8.*
